move unicode constants to edenutil
